chore: add environment example and login styles
