+++ .env.example
@@ -1,0 +1,5 @@
+PORT=3000
+JWT_SECRET=
+JWT_EXPIRES=24h
+PRINTER_NAME=EPSON TM-T20III Receipt
+NODE_ENV=development
