--- EPSON TM-T20IIII Receipt
@@ -1,34 +1,0 @@
-
-              🧪 TICKET DE PRUEBA
-==================================
-
-MI RESTAURANTE
-Av. Heroínas 123
-Cochabamba, Bolivia
-Tel: +591 4 XXX XXXX
-
-==================================
-
-📅 Fecha: 21/08/2025 23:45:50
-🖨️ Modelo: EPSON TM-T20IIII Receipt
-🔧 Método: Windows Spooler
-📊 Estado: ✅ FUNCIONANDO
-
-==================================
-
-Prueba de caracteres especiales:
-ñáéíóúü ÑÁÉÍÓÚÜ
-¡¿°±×÷§¶
-$ € £ ¥ ₹
-
-==================================
-
-        ✅ PRUEBA EXITOSA
-   La impresora funciona correctamente!
-   
-   Ahora puedes usar el sistema POS
-
-
-
-
-
