+++ README.md
@@ -1,1 +1,17 @@
 # POSALE
+
+## Setup
+
+1. Copia `.env.example` a `.env` y define los valores necesarios:
+   ```bash
+   cp .env.example .env
+   ```
+2. Instala dependencias:
+   ```bash
+   npm install
+   ```
+3. Inicia el servidor:
+   ```bash
+   npm start
+   ```
+
