+++ server/models/User.js
@@ -5,14 +5,13 @@ const bcrypt = require('bcryptjs');
 class User {
     static async create(userData) {
         await database.ensureConnected();
-        
+        const hashedPassword = await bcrypt.hash(userData.password, 10);
         return new Promise((resolve, reject) => {
-            const hashedPassword = bcrypt.hashSync(userData.password, 10);
             const sql = `
                 INSERT INTO users (username, password, full_name, role)
                 VALUES (?, ?, ?, ?)
             `;
-            
+
             database.getDB().run(
                 sql,
                 [userData.username, hashedPassword, userData.full_name, userData.role || 'user'],
@@ -26,10 +25,10 @@ class User {
 
     static async findByUsername(username) {
         await database.ensureConnected();
-        
+
         return new Promise((resolve, reject) => {
             const sql = `SELECT * FROM users WHERE username = ? AND active = 1`;
-            
+
             database.getDB().get(sql, [username], (err, row) => {
                 if (err) reject(err);
                 else resolve(row);
@@ -39,10 +38,10 @@ class User {
 
     static async findById(id) {
         await database.ensureConnected();
-        
+
         return new Promise((resolve, reject) => {
             const sql = `SELECT * FROM users WHERE id = ? AND active = 1`;
-            
+
             database.getDB().get(sql, [id], (err, row) => {
                 if (err) reject(err);
                 else resolve(row);
@@ -52,14 +51,14 @@ class User {
 
     static async findAll() {
         await database.ensureConnected();
-        
+
         return new Promise((resolve, reject) => {
             const sql = `
-                SELECT id, username, full_name, role, active, created_at 
-                FROM users 
+                SELECT id, username, full_name, role, active, created_at
+                FROM users
                 ORDER BY created_at DESC
             `;
-            
+
             database.getDB().all(sql, [], (err, rows) => {
                 if (err) reject(err);
                 else resolve(rows);
@@ -69,26 +68,25 @@ class User {
 
     static async update(id, userData) {
         await database.ensureConnected();
-        
-        return new Promise((resolve, reject) => {
+
+        return new Promise(async (resolve, reject) => {
             let sql = `
-                UPDATE users 
+                UPDATE users
                 SET username = ?, full_name = ?, role = ?, updated_at = CURRENT_TIMESTAMP
                 WHERE id = ?
             `;
             let params = [userData.username, userData.full_name, userData.role, id];
 
-            // Si se proporciona nueva contraseña, incluirla
             if (userData.password) {
-                const hashedPassword = bcrypt.hashSync(userData.password, 10);
+                const hashedPassword = await bcrypt.hash(userData.password, 10);
                 sql = `
-                    UPDATE users 
+                    UPDATE users
                     SET username = ?, password = ?, full_name = ?, role = ?, updated_at = CURRENT_TIMESTAMP
                     WHERE id = ?
                 `;
                 params = [userData.username, hashedPassword, userData.full_name, userData.role, id];
             }
-            
+
             database.getDB().run(sql, params, function(err) {
                 if (err) reject(err);
                 else resolve({ id, ...userData });
@@ -98,10 +96,10 @@ class User {
 
     static async delete(id) {
         await database.ensureConnected();
-        
+
         return new Promise((resolve, reject) => {
             const sql = `UPDATE users SET active = 0, updated_at = CURRENT_TIMESTAMP WHERE id = ?`;
-            
+
             database.getDB().run(sql, [id], function(err) {
                 if (err) reject(err);
                 else resolve({ deleted: true });
@@ -110,31 +108,29 @@ class User {
     }
 
     static async verifyPassword(plainPassword, hashedPassword) {
-        return bcrypt.compareSync(plainPassword, hashedPassword);
+        return bcrypt.compare(plainPassword, hashedPassword);
     }
 
     static async changePassword(id, oldPassword, newPassword) {
         await database.ensureConnected();
-        
-        // Primero verificar la contraseña actual
+
         const user = await this.findById(id);
         if (!user) {
             throw new Error('Usuario no encontrado');
         }
 
-        if (!this.verifyPassword(oldPassword, user.password)) {
+        if (!await this.verifyPassword(oldPassword, user.password)) {
             throw new Error('Contraseña actual incorrecta');
         }
 
-        // Actualizar con la nueva contraseña
+        const hashedPassword = await bcrypt.hash(newPassword, 10);
         return new Promise((resolve, reject) => {
-            const hashedPassword = bcrypt.hashSync(newPassword, 10);
             const sql = `
-                UPDATE users 
+                UPDATE users
                 SET password = ?, updated_at = CURRENT_TIMESTAMP
                 WHERE id = ?
             `;
-            
+
             database.getDB().run(sql, [hashedPassword, id], function(err) {
                 if (err) reject(err);
                 else resolve({ passwordChanged: true });
@@ -144,3 +140,4 @@ class User {
 }
 
 module.exports = User;
+
